requirements.txt
Flask==0.9
Jinja2==2.6
gunicorn==0.17.2
pymongo==2.6.3
requests==2.2.1
facebook-sdk==1.0.0